+++ src/components/Editor/EditorInput/EditorInput.js
@@ -92,7 +92,7 @@ const EditorInput = () => {
                     >
                       {fileName}
                       {!["main.py", "index.html"].includes(fileName) ?
-                        <Button className='btn--tertiary react-tabs__tab-inner-close-btn' label='close' onKeyDown={(e) => e.stopPropagation()} onClickHandler={(e) => closeFileTab(e, fileName)} ButtonIcon={() => <CloseIcon scaleFactor={0.85}/> }/>
+                        <Button className='btn--tertiary react-tabs__tab-inner-close-btn' label='close' onClickHandler={(e) => closeFileTab(e, fileName)} ButtonIcon={() => <CloseIcon scaleFactor={0.85}/> }/>
                       : null
                       }
                     </span>