+++ src/components/Editor/EditorSlice.js
@@ -104,6 +104,7 @@ export const EditorSlice = createSlice({
     notFoundModalShowing: false,
     newFileModalShowing: false,
     renameFileModalShowing: false,
+    newProjectModalShowing: false,
     renameProjectModalShowing: false,
     deleteProjectModalShowing: false,
     modals: {},
@@ -283,6 +284,12 @@ export const EditorSlice = createSlice({
       state.renameFileModalShowing = false;
       state.nameError = "";
     },
+    showNewProjectModal: (state) => {
+      state.newProjectModalShowing = true
+    },
+    closeNewProjectModal: (state) => {
+      state.newProjectModalShowing = false
+    },
     showRenameProjectModal: (state, action) => {
       state.modals.renameProject = action.payload;
       state.renameProjectModalShowing = true;
@@ -455,6 +462,8 @@ export const {
   closeNewFileModal,
   showRenameFileModal,
   closeRenameFileModal,
+  showNewProjectModal,
+  closeNewProjectModal,
   showRenameProjectModal,
   closeRenameProjectModal,
   showDeleteProjectModal,