+++ src/components/Editor/Runners/HtmlRunner/HtmlRunner.test.js
@@ -19,6 +19,7 @@ const allowedLinkHTMLPage = {
   name: "allowed_link",
   extension: "html",
   content: '<head></head><body><a href="#">ANCHOR LINK!</a></body>',
+
 };
 
 describe("When page first loaded", () => {
@@ -128,6 +129,7 @@ describe("When page first loaded in embedded viewer", () => {
 });
 
 describe("When run run triggered", () => {
+
   let store;
 
   beforeEach(() => {