+++ src/components/Modals/RenameFile.test.js
@@ -1,5 +1,5 @@
 import React from "react";
-import { fireEvent, render } from "@testing-library/react"
+import { fireEvent, render, screen } from "@testing-library/react"
 import { Provider } from 'react-redux';
 import configureStore from 'redux-mock-store';
 
@@ -10,7 +10,6 @@ describe("Testing the rename file modal", () => {
   let store;
   let inputBox;
   let saveButton;
-  let getByText;
 
   beforeEach(() => {
     const middlewares = []
@@ -42,18 +41,17 @@ describe("Testing the rename file modal", () => {
       }
     }
     store = mockStore(initialState);
-    ({getByText} = render(<Provider store={store}><div id='app'><RenameFile currentName='main' currentExtension='py' fileKey={0} /></div></Provider>))
-    inputBox = document.getElementById('name')
-    saveButton = getByText('filePane.renameFileModal.save');
+    render(<Provider store={store}><div id='app'><RenameFile/></div></Provider>)
+    inputBox = screen.getByRole('textbox')
+    saveButton = screen.getByText('filePane.renameFileModal.save').closest('button');
   })
 
   test('State being set displays the modal', () => {
-    expect(getByText('filePane.renameFileModal.heading')).toBeInTheDocument()
+    expect(screen.getByText('filePane.renameFileModal.heading')).toBeInTheDocument()
   })
 
   test("Pressing save renames the file to the given name", () => {
     fireEvent.change(inputBox, {target: {value: "file1.py"}})
-    inputBox.innerHTML = "file1.py";
     fireEvent.click(saveButton)
     const expectedActions = [
       updateComponentName({key: 0, extension: "py", name: "file1"}),
@@ -64,7 +62,6 @@ describe("Testing the rename file modal", () => {
 
   test("Pressing Enter renames the file to the given name", () => {
     fireEvent.change(inputBox, {target: {value: "file1.py"}})
-    inputBox.innerHTML = "file1.py";
     fireEvent.keyDown(inputBox, { key: 'Enter'})
     const expectedActions = [
       updateComponentName({key: 0, extension: "py", name: "file1"}),