+++ src/i18n.js
@@ -238,6 +238,21 @@ i18n
           modals: {
             close: 'Close'
           },
+          newProjectModal: {
+            cancel: 'Cancel',
+            createProject: 'Create project',
+            heading: 'Create a new project',
+            projectName: {
+              default: 'Untitled',
+              helpText: 'You can always rename your project later',
+              inputLabel: 'Project name',
+            },
+            projectType: {
+              html: 'HTML',
+              inputLabel: 'What kind of project do you want to make?',
+              python: 'Python'
+            }
+          },
           notifications: {
             close: "close",
             loginPrompt: