+++ CHANGELOG.md
@@ -8,9 +8,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
+- Ability to specify project name and type on creation (#519)
+
 ### Changed
 
 ### Fixed
+
+- Input modal help text styling in light mode (#519)
+- Propagation of key press events in modals (#519)
 
 ## [0.16.1] - 2023-05-19